Backend/requirements.txt
certifi==2019.3.9
pipenv==2018.11.26
virtualenv==16.6.0
virtualenv-clone==0.5.3
django-rest-framework
django-cors-headers
django-rest-framework-mongoengine
mongoengine
djongo